Dealing with original zip code issue
It turns out that the zip code tabulation areas used by the US Census do not match one-to-one with zip codes in the dataset. I replaced the original plot with one points with the geographic centroid of each zip code (some of which had to be manually collected with Google Maps).