(Very) successful code optimization!
The ABM with angles is down from 2.65 minutes to 13.6 seconds. I think it would be quite difficult to get runtime down any further. I tried to leave the unique brands and component/angle combinations as integers in matrices rather than converting to strings, and the string version is actually significantly faster.